saas/backend/requirements.txt
# ── Framework Web ─────────────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1
pydantic[email]==2.7.1

# ── Base de données ───────────────────────────────────────────────────────────
sqlalchemy==2.0.30
asyncpg==0.29.0          # Driver PostgreSQL async
psycopg2-binary==2.9.9   # Driver PostgreSQL sync (Celery workers)
alembic==1.13.1          # Migrations

# ── Géospatial ────────────────────────────────────────────────────────────────
gdal==3.8.4              # GDAL/OGR (version doit matcher le système)
geopandas==0.14.3
fiona==1.9.5
shapely==2.0.3
pyproj==3.6.1

# ── Workers asynchrones ───────────────────────────────────────────────────────
celery==5.3.6
redis==5.0.4
flower==2.0.1            # Dashboard monitoring Celery

# ── Sécurité ─────────────────────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9  # Upload form-data

# ── Stockage cloud ────────────────────────────────────────────────────────────
boto3==1.34.87

# ── Paiement ─────────────────────────────────────────────────────────────────
stripe==9.3.0

# ── Monitoring ───────────────────────────────────────────────────────────────
sentry-sdk[fastapi]==1.45.0
prometheus-fastapi-instrumentator==7.0.0

# ── Utilitaires ───────────────────────────────────────────────────────────────
pandas==2.2.0
numpy==1.26.4
python-dotenv==1.0.1
httpx==0.27.0            # Client HTTP async (tests)
